feat: cross-platform terminal with ConPTY on Windows, PTY on Unix
Split terminal handler into platform-specific files:
- terminal_windows.go: uses ConPTY for full interactive shell
  (character echo, arrow keys, tab completion, resize support)
- terminal_unix.go: uses creack/pty (existing behavior, unchanged)

Windows defaults to cmd.exe (COMSPEC), Unix to $SHELL or /bin/sh.
Both provide identical interactive terminal experience in the web UI.
