+++ README.md
@@ -17,7 +17,7 @@ This project is a fork of the excellent [tcarac/taskboard](https://github.com/tc
 - **GitHub Sync Engine** — Bidirectional sync between local SQLite and GitHub Issues. Metadata is stored in hidden HTML comments.
 - **Strict Mode** — Enforce professional standards. When enabled, tickets *must* have a User Story and Acceptance Criteria (Gherkin) before they can be created or updated.
 - **Git Hooks** — Automated sync on `git push` and `git pull` via installed `pre-push` and `post-merge` hooks.
-- **Embedded Terminal** — Run AI coding agents directly from the web UI.
+- **Embedded Terminal** — Full interactive shell in the web UI. Uses ConPTY on Windows (PowerShell/cmd.exe) and PTY on macOS/Linux (bash/zsh). Run commands, AI agents, or debug directly from the board.
 
 ## Requirements
