+++ internal/server/server.go
@@ -8,12 +8,8 @@ import (
 	"log/slog"
 	"net"
 	"net/http"
-	"os"
-	"os/exec"
-	"sync"
 	"time"
 
-	"github.com/creack/pty"
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/go-chi/cors"
@@ -731,76 +727,6 @@ func (s *Server) getBoard(w http.ResponseWriter, r *http.Request) {
 
 var wsUpgrader = websocket.Upgrader{
 	CheckOrigin: func(r *http.Request) bool { return true },
-}
-
-func (s *Server) handleTerminalWS(w http.ResponseWriter, r *http.Request) {
-	conn, err := wsUpgrader.Upgrade(w, r, nil)
-	if err != nil {
-		return
-	}
-	defer conn.Close()
-
-	shell := os.Getenv("SHELL")
-	if shell == "" {
-		shell = "/bin/sh"
-	}
-	// #nosec G204 G702
-	cmd := exec.Command(shell)
-	cmd.Env = append(os.Environ(), "TERM=xterm-256color")
-
-	ptmx, err := pty.Start(cmd)
-	if err != nil {
-		_ = conn.WriteMessage(websocket.TextMessage, []byte(`{"error":"failed to start terminal"}`))
-		return
-	}
-	defer ptmx.Close()
-
-	var once sync.Once
-	cleanup := func() {
-		once.Do(func() {
-			_ = cmd.Process.Kill()
-			_ = cmd.Wait()
-		})
-	}
-	defer cleanup()
-
-	// PTY → WebSocket
-	go func() {
-		buf := make([]byte, 4096)
-		for {
-			n, err := ptmx.Read(buf)
-			if err != nil {
-				_ = conn.Close()
-				return
-			}
-			if err := conn.WriteMessage(websocket.BinaryMessage, buf[:n]); err != nil {
-				return
-			}
-		}
-	}()
-
-	// WebSocket → PTY
-	for {
-		msgType, msg, err := conn.ReadMessage()
-		if err != nil {
-			cleanup()
-			return
-		}
-
-		switch msgType {
-		case websocket.BinaryMessage:
-			_, _ = ptmx.Write(msg)
-		case websocket.TextMessage:
-			var ctrl struct {
-				Type string `json:"type"`
-				Cols uint16 `json:"cols"`
-				Rows uint16 `json:"rows"`
-			}
-			if json.Unmarshal(msg, &ctrl) == nil && ctrl.Type == "resize" {
-				_ = pty.Setsize(ptmx, &pty.Winsize{Cols: ctrl.Cols, Rows: ctrl.Rows})
-			}
-		}
-	}
 }
 
 func (s *Server) healthCheck(w http.ResponseWriter, r *http.Request) {
